+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refined oneDrive rate limiter controls to reduce throttling errors.
 - Fix handling of duplicate folders at the same hierarchy level in Exchange. Duplicate folders will be merged during restore operations.
 - Fix backup for mailboxes that has used up all their storage quota
+- Restored folders no longer appear in the Restore results. Only restored items will be displayed.
 
 ### Known Issues
 - Restore operations will merge duplicate Exchange folders at the same hierarchy level into a single folder.